moved engine.prepare to its own file
